requirements.txt
pandas
numpy
requests
tqdm
python-dotenv